src/main/java/com/ing/demos/web/service/ProductService.java
package com.ing.demos.web.service;

import com.ing.demos.persistence.model.ProductEntity;
import com.ing.demos.persistence.repository.ProductRepository;
import com.ing.demos.web.exception.EntityNotFoundException;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class ProductService {

    private final ProductRepository productRepository;

    public ProductService(ProductRepository productRepository) {
        this.productRepository = productRepository;
    }

    public List<ProductEntity> findAllProducts() {
        return (List<ProductEntity>) productRepository.findAll();
    }

    public ProductEntity findProductById(Long id) {
        return productRepository.findById(id)
                .orElseThrow(EntityNotFoundException::new);
    }

    public ProductEntity saveProduct(ProductEntity product) {
        return productRepository.save(product);
    }

    public void deleteProduct(Long id) {
        productRepository.deleteById(id);
    }
}
